quitando algunos warnings y arreglado falla en un estado de player
ahora si estas realizando un ataque y te golpean, se cancela el ataque

+++ src/MF-FighterTraighter/MF-FighterTraighter/WindowManager.h
@@ -33,7 +33,7 @@ public:
 	inline int getCurResolutionIndex() { return currentResolution_; }
 
 	//get the number of resolutions that are supported in the computer 
-	inline int getAvailableResolutions() { return supportedResolutions_.size(); }
+	inline size_t getAvailableResolutions() { return supportedResolutions_.size(); }
 
 	//returns true if it's the fullscreen on
 	inline bool getFullscreen() { return fullscreen_; }
